fix:marketDetail 메인 화면에서 data fetch함수 수정
